requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
asttokens==2.2.1
    # via stack-data
backcall==0.2.0
    # via ipython
click==8.1.3
    # via flask
decorator==5.1.1
    # via ipython
executing==1.2.0
    # via stack-data
flask==2.2.2
    # via -r requirements.in
hurry-filesize==0.9
    # via -r requirements.in
ipython==8.7.0
    # via -r requirements.in
itsdangerous==2.1.2
    # via flask
jedi==0.18.2
    # via ipython
jinja2==3.1.2
    # via flask
markupsafe==2.1.1
    # via
    #   jinja2
    #   werkzeug
matplotlib-inline==0.1.6
    # via ipython
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
prompt-toolkit==3.0.36
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pygments==2.13.0
    # via ipython
six==1.16.0
    # via asttokens
stack-data==0.6.2
    # via ipython
traitlets==5.7.0
    # via
    #   ipython
    #   matplotlib-inline
wcwidth==0.2.5
    # via prompt-toolkit
werkzeug==2.2.2
    # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
